chore(medusa,types): [9] Add request types to API routes
