+++ packages/core/types/src/http/order/admin/queries.ts
@@ -1,0 +1,3 @@
+import { BaseOrderFilters } from "../common";
+
+export interface AdminOrderFilters extends BaseOrderFilters {}
