+++ packages/core/types/src/http/order/admin/responses.ts
@@ -1,0 +1,14 @@
+import { PaginatedResponse } from "../../common"
+import { AdminOrder, AdminOrderPreview } from "./entities"
+
+export interface AdminOrderResponse {
+  order: AdminOrder
+}
+
+export type AdminOrderListResponse = PaginatedResponse<{
+  orders: AdminOrder[]
+}>
+
+export interface AdminOrderPreviewResponse {
+  order: AdminOrderPreview
+}
