+++ packages/medusa/src/api/admin/payments/route.ts
@@ -7,10 +7,11 @@ import {
   MedusaResponse,
 } from "../../../types/routing"
 import { AdminGetPaymentsParamsType } from "./validators"
+import { HttpTypes } from "@medusajs/types"
 
 export const GET = async (
   req: AuthenticatedMedusaRequest<AdminGetPaymentsParamsType>,
-  res: MedusaResponse
+  res: MedusaResponse<HttpTypes.AdminPaymentsResponse>
 ) => {
   const remoteQuery = req.scope.resolve(ContainerRegistrationKeys.REMOTE_QUERY)
   const queryObject = remoteQueryObjectFromString({
